requirements.txt
pyserial>=3.5
watchdog>=2.1.0
